+++ src/main/java/io/github/astrapi69/swing/mouse/MouseClickType.java
@@ -1,6 +1,0 @@
-package io.github.astrapi69.swing.mouse;
-
-public enum MouseClickType
-{
-	SINGLE_LEFT, DOUBLE_LEFT, SINGLE_MIDDLE, DOUBLE_MIDDLE, SINGLE_RIGHT, DOUBLE_RIGHT
-}